requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
joblib>=1.3.0
Pillow>=9.5.0
